Abstract building into schematic provider and requesting building.
